Extract OpportunityDetailPage and OpportunityOverviewPage components
